Zipper - Split RadosStore from base SAL
