+++ src/rgw/rgw_dencoder.cc
@@ -2,6 +2,7 @@
 // vim: ts=8 sw=2 smarttab ft=cpp
 
 #include "rgw_common.h"
+#include "rgw_rados.h"
 #include "rgw_zone.h"
 #include "rgw_log.h"
 #include "rgw_acl.h"
