+++ src/rgw/rgw_gc.h
@@ -12,6 +12,7 @@
 #include "common/Thread.h"
 #include "rgw_common.h"
 #include "rgw_sal.h"
+#include "rgw_rados.h"
 #include "cls/rgw/cls_rgw_types.h"
 
 #include <atomic>
