+++ src/rgw/rgw_period_pusher.cc
@@ -8,6 +8,7 @@
 #include "rgw_cr_rest.h"
 #include "rgw_zone.h"
 #include "rgw_sal.h"
+#include "rgw_sal_rados.h"
 
 #include "services/svc_zone.h"
