+++ src/rgw/rgw_putobj_processor.cc
@@ -18,6 +18,7 @@
 #include "rgw_multi.h"
 #include "rgw_compression.h"
 #include "services/svc_sys_obj.h"
+#include "rgw_sal_rados.h"
 
 #define dout_subsys ceph_subsys_rgw
