+++ src/rgw/rgw_quota.cc
@@ -22,6 +22,7 @@
 
 #include "rgw_common.h"
 #include "rgw_sal.h"
+#include "rgw_sal_rados.h"
 #include "rgw_quota.h"
 #include "rgw_bucket.h"
 #include "rgw_user.h"
