+++ src/rgw/rgw_rest_bucket.cc
@@ -4,6 +4,7 @@
 #include "rgw_op.h"
 #include "rgw_bucket.h"
 #include "rgw_rest_bucket.h"
+#include "rgw_sal_rados.h"
 
 #include "include/str_list.h"
