+++ src/rgw/rgw_rest_config.cc
@@ -21,6 +21,7 @@
 #include "rgw_rest_s3.h"
 #include "rgw_rest_config.h"
 #include "rgw_client_io.h"
+#include "rgw_sal_rados.h"
 #include "common/errno.h"
 #include "include/ceph_assert.h"
