+++ src/rgw/rgw_rest_user_policy.cc
@@ -15,6 +15,7 @@
 #include "rgw_op.h"
 #include "rgw_rest.h"
 #include "rgw_rest_user_policy.h"
+#include "rgw_sal_rados.h"
 #include "services/svc_zone.h"
 
 #define dout_subsys ceph_subsys_rgw
