+++ src/rgw/rgw_sal.h
@@ -15,8 +15,12 @@
 
 #pragma once
 
-#include "rgw_rados.h"
 #include "rgw_user.h"
+#include "rgw_obj_manifest.h"
+
+class RGWGetDataCB;
+struct RGWObjState;
+class RGWAccessListFilter;
 
 namespace rgw { namespace sal {
 
@@ -443,280 +447,5 @@ class RGWObject {
     }
 };
 
-
-class RGWRadosStore;
-
-class RGWRadosUser : public RGWUser {
-  private:
-    RGWRadosStore *store;
-
-  public:
-    RGWRadosUser(RGWRadosStore *_st, const rgw_user& _u) : RGWUser(_u), store(_st) { }
-    RGWRadosUser(RGWRadosStore *_st, const RGWUserInfo& _i) : RGWUser(_i), store(_st) { }
-    RGWRadosUser(RGWRadosStore *_st) : store(_st) { }
-    RGWRadosUser() {}
-
-    int list_buckets(const std::string& marker, const std::string& end_marker,
-				uint64_t max, bool need_stats, RGWBucketList& buckets);
-    RGWBucket* create_bucket(rgw_bucket& bucket, ceph::real_time creation_time);
-
-    /* Placeholders */
-    virtual int load_by_id(optional_yield y);
-
-    friend class RGWRadosBucket;
-};
-
-class RGWRadosObject : public RGWObject {
-  private:
-    RGWRadosStore *store;
-    RGWAccessControlPolicy acls;
-
-  public:
-
-    struct RadosReadOp : public ReadOp {
-    private:
-      RGWRadosObject* source;
-      RGWObjectCtx* rctx;
-      RGWRados::Object op_target;
-      RGWRados::Object::Read parent_op;
-
-    public:
-      RadosReadOp(RGWRadosObject *_source, RGWObjectCtx *_rctx);
-
-      virtual int prepare(optional_yield y) override;
-      virtual int read(int64_t ofs, int64_t end, bufferlist& bl, optional_yield y) override;
-      virtual int iterate(int64_t ofs, int64_t end, RGWGetDataCB *cb, optional_yield y) override;
-      virtual int get_manifest(RGWObjManifest **pmanifest, optional_yield y) override;
-    };
-
-    RGWRadosObject() = default;
-
-    RGWRadosObject(RGWRadosStore *_st, const rgw_obj_key& _k)
-      : RGWObject(_k),
-	store(_st),
-        acls() {
-    }
-    RGWRadosObject(RGWRadosStore *_st, const rgw_obj_key& _k, RGWBucket* _b)
-      : RGWObject(_k, _b),
-	store(_st),
-        acls() {
-    }
-    RGWRadosObject(RGWRadosObject& _o) = default;
-
-    int read(off_t offset, off_t length, std::iostream& stream) { return length; }
-    int write(off_t offset, off_t length, std::iostream& stream) { return length; }
-    virtual int delete_object(RGWObjectCtx* obj_ctx, ACLOwner obj_owner,
-			      ACLOwner bucket_owner, ceph::real_time unmod_since,
-			      bool high_precision_time, uint64_t epoch,
-			      std::string& version_id,optional_yield y) override;
-    virtual int copy_object(RGWObjectCtx& obj_ctx, RGWUser* user,
-               req_info *info, const rgw_zone_id& source_zone,
-               rgw::sal::RGWObject* dest_object, rgw::sal::RGWBucket* dest_bucket,
-               rgw::sal::RGWBucket* src_bucket,
-               const rgw_placement_rule& dest_placement,
-               ceph::real_time *src_mtime, ceph::real_time *mtime,
-               const ceph::real_time *mod_ptr, const ceph::real_time *unmod_ptr,
-               bool high_precision_time,
-               const char *if_match, const char *if_nomatch,
-               AttrsMod attrs_mod, bool copy_if_newer, RGWAttrs& attrs,
-               RGWObjCategory category, uint64_t olh_epoch,
-	       boost::optional<ceph::real_time> delete_at,
-               string *version_id, string *tag, string *etag,
-               void (*progress_cb)(off_t, void *), void *progress_data,
-               const DoutPrefixProvider *dpp, optional_yield y) override;
-    RGWAccessControlPolicy& get_acl(void) { return acls; }
-    int set_acl(const RGWAccessControlPolicy& acl) { acls = acl; return 0; }
-    virtual void set_atomic(RGWObjectCtx *rctx) const;
-    virtual void set_prefetch_data(RGWObjectCtx *rctx);
-
-    virtual int get_obj_state(RGWObjectCtx *rctx, RGWBucket& bucket, RGWObjState **state, optional_yield y, bool follow_olh = true) override;
-    virtual int set_obj_attrs(RGWObjectCtx* rctx, RGWAttrs* setattrs, RGWAttrs* delattrs, optional_yield y, rgw_obj* target_obj = NULL) override;
-    virtual int get_obj_attrs(RGWObjectCtx *rctx, optional_yield y, rgw_obj* target_obj = NULL) override;
-    virtual int modify_obj_attrs(RGWObjectCtx *rctx, const char *attr_name, bufferlist& attr_val, optional_yield y) override;
-    virtual int delete_obj_attrs(RGWObjectCtx *rctx, const char *attr_name, optional_yield y) override;
-    virtual int copy_obj_data(RGWObjectCtx& rctx, RGWBucket* dest_bucket, RGWObject* dest_obj, uint16_t olh_epoch, std::string* petag, const DoutPrefixProvider *dpp, optional_yield y) override;
-    virtual bool is_expired() override;
-    virtual void gen_rand_obj_instance_name() override;
-    virtual std::unique_ptr<RGWObject> clone() {
-      return std::unique_ptr<RGWObject>(new RGWRadosObject(*this));
-    }
-
-    /* OPs */
-    virtual std::unique_ptr<ReadOp> get_read_op(RGWObjectCtx *) override;
-
-    /* OMAP */
-    virtual int omap_get_vals_by_keys(const std::string& oid,
-			      const std::set<std::string>& keys,
-			      std::map<std::string, bufferlist> *vals) override;
-
-  private:
-    int read_attrs(RGWRados::Object::Read &read_op, optional_yield y, rgw_obj *target_obj = nullptr);
-};
-
-class RGWRadosBucket : public RGWBucket {
-  private:
-    RGWRadosStore *store;
-    RGWAccessControlPolicy acls;
-
-  public:
-    RGWRadosBucket(RGWRadosStore *_st)
-      : store(_st),
-        acls() {
-    }
-
-    RGWRadosBucket(RGWRadosStore *_st, const rgw_bucket& _b)
-      : RGWBucket(_b),
-	store(_st),
-        acls() {
-    }
-
-    RGWRadosBucket(RGWRadosStore *_st, const RGWBucketEnt& _e)
-      : RGWBucket(_e),
-	store(_st),
-        acls() {
-    }
-
-    RGWRadosBucket(RGWRadosStore *_st, const RGWBucketInfo& _i)
-      : RGWBucket(_i),
-	store(_st),
-        acls() {
-    }
-
-    RGWRadosBucket(RGWRadosStore *_st, const rgw_bucket& _b, RGWUser* _u)
-      : RGWBucket(_b, _u),
-	store(_st),
-        acls() {
-    }
-
-    RGWRadosBucket(RGWRadosStore *_st, const RGWBucketEnt& _e, RGWUser* _u)
-      : RGWBucket(_e, _u),
-	store(_st),
-        acls() {
-    }
-
-    RGWRadosBucket(RGWRadosStore *_st, const RGWBucketInfo& _i, RGWUser* _u)
-      : RGWBucket(_i, _u),
-	store(_st),
-        acls() {
-    }
-
-    ~RGWRadosBucket() { }
-
-    virtual int load_by_name(const std::string& tenant, const std::string& bucket_name, const std::string bucket_instance_id, RGWSysObjectCtx *rctx, optional_yield y) override;
-    virtual std::unique_ptr<RGWObject> get_object(const rgw_obj_key& k) override;
-    RGWBucketList* list(void) { return new RGWBucketList(); }
-    virtual int list(ListParams&, int, ListResults&, optional_yield y) override;
-    RGWObject* create_object(const rgw_obj_key& key /* Attributes */) override;
-    virtual int remove_bucket(bool delete_children, std::string prefix, std::string delimiter, bool forward_to_master, req_info* req_info, optional_yield y) override;
-    RGWAccessControlPolicy& get_acl(void) { return acls; }
-    virtual int set_acl(RGWAccessControlPolicy& acl, optional_yield y) override;
-    virtual int get_bucket_info(optional_yield y) override;
-    virtual int get_bucket_stats(RGWBucketInfo& bucket_info, int shard_id,
-				 std::string *bucket_ver, std::string *master_ver,
-				 std::map<RGWObjCategory, RGWStorageStats>& stats,
-				 std::string *max_marker = nullptr,
-				 bool *syncstopped = nullptr) override;
-    virtual int read_bucket_stats(optional_yield y) override;
-    virtual int sync_user_stats() override;
-    virtual int update_container_stats(void) override;
-    virtual int check_bucket_shards(void) override;
-    virtual int link(RGWUser* new_user, optional_yield y) override;
-    virtual int unlink(RGWUser* new_user, optional_yield y) override;
-    virtual int chown(RGWUser* new_user, RGWUser* old_user, optional_yield y) override;
-    virtual int put_instance_info(bool exclusive, ceph::real_time mtime) override;
-    virtual bool is_owner(RGWUser* user) override;
-    virtual int check_empty(optional_yield y) override;
-    virtual int check_quota(RGWQuotaInfo& user_quota, RGWQuotaInfo& bucket_quota, uint64_t obj_size, bool check_size_only = false) override;
-    virtual int set_instance_attrs(RGWAttrs& attrs, optional_yield y) override;
-    virtual int try_refresh_info(ceph::real_time *pmtime) override;
-    virtual std::unique_ptr<RGWBucket> clone() {
-      return std::unique_ptr<RGWBucket>(new RGWRadosBucket(*this));
-    }
-
-    friend class RGWRadosStore;
-};
-
-class RGWRadosStore : public RGWStore {
-  private:
-    RGWRados *rados;
-    RGWUserCtl *user_ctl;
-
-  public:
-    RGWRadosStore()
-      : rados(nullptr) {
-      }
-    ~RGWRadosStore() {
-      delete rados;
-    }
-
-    virtual std::unique_ptr<RGWUser> get_user(const rgw_user& u);
-    virtual std::unique_ptr<RGWObject> get_object(const rgw_obj_key& k) override;
-    virtual int get_bucket(RGWUser* u, const rgw_bucket& b, std::unique_ptr<RGWBucket>* bucket) override;
-    virtual int get_bucket(RGWUser* u, const RGWBucketInfo& i, std::unique_ptr<RGWBucket>* bucket) override;
-    virtual int get_bucket(RGWUser* u, const std::string& tenant, const std::string&name, std::unique_ptr<RGWBucket>* bucket) override;
-    virtual int create_bucket(RGWUser& u, const rgw_bucket& b,
-                            const std::string& zonegroup_id,
-                            rgw_placement_rule& placement_rule,
-                            std::string& swift_ver_location,
-                            const RGWQuotaInfo * pquota_info,
-			    map<std::string, bufferlist>& attrs,
-                            RGWBucketInfo& info,
-                            obj_version& ep_objv,
-			    bool exclusive,
-			    bool obj_lock_enabled,
-			    bool *existed,
-			    req_info& req_info,
-			    std::unique_ptr<RGWBucket>* bucket);
-    virtual RGWBucketList* list_buckets(void) { return new RGWBucketList(); }
-    virtual bool is_meta_master() override;
-    virtual int forward_request_to_master(RGWUser* user, obj_version *objv,
-				  bufferlist& in_data, JSONParser *jp, req_info& info) override;
-    virtual int defer_gc(RGWObjectCtx *rctx, RGWBucket* bucket, RGWObject* obj,
-			 optional_yield y) override;
-
-    void setRados(RGWRados * st) { rados = st; }
-    RGWRados *getRados(void) { return rados; }
-
-    RGWServices *svc() { return &rados->svc; }
-    const RGWServices *svc() const { return &rados->svc; }
-    RGWCtl *ctl() { return &rados->ctl; }
-    const RGWCtl *ctl() const { return &rados->ctl; }
-
-    void setUserCtl(RGWUserCtl *_ctl) { user_ctl = _ctl; }
-
-    void finalize(void) override;
-
-    virtual CephContext *ctx(void) { return rados->ctx(); }
-
-
-    int get_obj_head_ioctx(const RGWBucketInfo& bucket_info, const rgw_obj& obj,
-			   librados::IoCtx *ioctx);
-
-    // implements DoutPrefixProvider
-    std::ostream& gen_prefix(std::ostream& out) const { return out << "RGWRadosStore "; }
-    CephContext* get_cct() const override { return rados->ctx(); }
-    unsigned get_subsys() const override { return ceph_subsys_rgw; }
-
-};
-
 } } // namespace rgw::sal
 
-
-class RGWStoreManager {
-public:
-  RGWStoreManager() {}
-  static rgw::sal::RGWRadosStore *get_storage(CephContext *cct, bool use_gc_thread, bool use_lc_thread, bool quota_threads,
-			       bool run_sync_thread, bool run_reshard_thread, bool use_cache = true) {
-    rgw::sal::RGWRadosStore *store = init_storage_provider(cct, use_gc_thread, use_lc_thread,
-	quota_threads, run_sync_thread, run_reshard_thread, use_cache);
-    return store;
-  }
-  static rgw::sal::RGWRadosStore *get_raw_storage(CephContext *cct) {
-    rgw::sal::RGWRadosStore *rados = init_raw_storage_provider(cct);
-    return rados;
-  }
-  static rgw::sal::RGWRadosStore *init_storage_provider(CephContext *cct, bool use_gc_thread, bool use_lc_thread, bool quota_threads, bool run_sync_thread, bool run_reshard_thread, bool use_metadata_cache);
-  static rgw::sal::RGWRadosStore *init_raw_storage_provider(CephContext *cct);
-  static void close_storage(rgw::sal::RGWRadosStore *store);
-
-};
